Feat: Complete Dashboard overhaul (Bento Grid), fixed BigDecimal/Date logic, and restored Charts

+++ src/main/java/com/example/demo/model/Customer.java
@@ -6,19 +6,22 @@ import jakarta.persistence.*;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id; // 👈 FIXED: Changed Integer to Long
 
     private String phone;
     private String name;
     private Integer points;
 
     // Getters and Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; } // 👈 Updated return type
+    public void setId(Long id) { this.id = id; } // 👈 Updated parameter type
+    
     public String getPhone() { return phone; }
     public void setPhone(String phone) { this.phone = phone; }
+    
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
+    
     public Integer getPoints() { return points; }
     public void setPoints(Integer points) { this.points = points; }
 }
